chore: rename locker
